backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose==3.3.0
python-dotenv==1.0.0
firebase-admin==6.2.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
pandas==2.1.1
numpy==1.25.2
Pillow==10.0.1
requests==2.31.0
difflib
uuid
base64